requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
pandas==2.2.2
openpyxl==3.1.2
openai>=1.37.0
numpy>=1.26.4
tenacity>=8.2.3
tqdm>=4.66.4
scikit-learn>=1.4.2
PyPDF2>=3.0.0
python-docx>=0.8.11
